+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 - Reduce delay copyable
+- Add javascript helpers on install
 
 ## [0.0.38] - 2024-09-24
 - Add local_time_controller